+++ Source/Device.generated.swift
@@ -202,22 +202,38 @@ public enum Device {
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP875/sp875-sp876-iphone14-pro-promax_2x.png)
     case iPhone14ProMax
-    /// Device is an [iPhone 15]()
+    /// Device is an [iPhone 15](https://support.apple.com/en-us/111831)
     ///
     /// ![Image]()
     case iPhone15
-    /// Device is an [iPhone 15 Plus]()
+    /// Device is an [iPhone 15 Plus](https://support.apple.com/en-us/111830)
     ///
     /// ![Image]()
     case iPhone15Plus
-    /// Device is an [iPhone 15 Pro]()
+    /// Device is an [iPhone 15 Pro](https://support.apple.com/en-us/111829)
     ///
     /// ![Image]()
     case iPhone15Pro
-    /// Device is an [iPhone 15 Pro Max]()
+    /// Device is an [iPhone 15 Pro Max](https://support.apple.com/en-us/111828)
     ///
     /// ![Image]()
     case iPhone15ProMax
+    /// Device is an [iPhone 16]()
+    ///
+    /// ![Image]()
+    case iPhone16
+    /// Device is an [iPhone 16 Plus]()
+    ///
+    /// ![Image]()
+    case iPhone16Plus
+    /// Device is an [iPhone 16 Pro]()
+    ///
+    /// ![Image]()
+    case iPhone16Pro
+    /// Device is an [iPhone 16 Pro Max]()
+    ///
+    /// ![Image]()
+    case iPhone16ProMax
     /// Device is an [iPad 2](https://support.apple.com/kb/SP622)
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP622/SP622_01-ipad2-mul.png)
@@ -274,6 +290,14 @@ public enum Device {
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP866/sp866-ipad-air-5gen_2x.png)
     case iPadAir5
+    /// Device is an [iPad Air 11-inch (M2)](https://support.apple.com/en-us/119894)
+    ///
+    /// ![Image](https://cdsassets.apple.com/content/services/pub/image?productid=301027&size=240x240)
+    case iPadAir11M2
+    /// Device is an [iPad Air 13-inch (M2)](https://support.apple.com/en-us/119893)
+    ///
+    /// ![Image](https://cdsassets.apple.com/content/services/pub/image?productid=301029&size=240x240)
+    case iPadAir13M2
     /// Device is an [iPad Mini](https://support.apple.com/kb/SP661)
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP661/sp661_ipad_mini_color.jpg)
@@ -346,6 +370,14 @@ public enum Device {
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP882/ipad-pro-4gen-mainimage_2x.png)
     case iPadPro12Inch6
+    /// Device is an [iPad Pro 11-inch (M4)](https://support.apple.com/en-us/119892)
+    ///
+    /// ![Image](https://cdsassets.apple.com/content/services/pub/image?productid=301031&size=240x240)
+    case iPadPro11M4
+    /// Device is an [iPad Pro 13-inch (M4)](https://support.apple.com/en-us/119891)
+    ///
+    /// ![Image](https://cdsassets.apple.com/content/services/pub/image?productid=301033&size=240x240)
+    case iPadPro13M4
     /// Device is a [HomePod](https://support.apple.com/kb/SP773)
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP773/homepod_space_gray_large_2x.jpg)
@@ -472,6 +504,14 @@ public enum Device {
     ///
     /// ![Image]()
     case appleWatchUltra2
+    /// Device is an [Apple Watch Series 10]()
+    ///
+    /// ![Image]()
+    case appleWatchSeries10_42mm
+    /// Device is an [Apple Watch Series 10]()
+    ///
+    /// ![Image]()
+    case appleWatchSeries10_46mm
   #elseif os(visionOS)
     case visionPro
   #endif
@@ -554,6 +594,10 @@ public enum Device {
       case "iPhone15,5": return iPhone15Plus
       case "iPhone16,1": return iPhone15Pro
       case "iPhone16,2": return iPhone15ProMax
+      case "iPhone17,3": return iPhone16
+      case "iPhone17,4": return iPhone16Plus
+      case "iPhone17,1": return iPhone16Pro
+      case "iPhone17,2": return iPhone16ProMax
       case "iPad2,1", "iPad2,2", "iPad2,3", "iPad2,4": return iPad2
       case "iPad3,1", "iPad3,2", "iPad3,3": return iPad3
       case "iPad3,4", "iPad3,5", "iPad3,6": return iPad4
@@ -568,6 +612,8 @@ public enum Device {
       case "iPad13,18", "iPad13,19": return iPad10
       case "iPad13,1", "iPad13,2": return iPadAir4
       case "iPad13,16", "iPad13,17": return iPadAir5
+      case "iPad14,8", "iPad14,9": return iPadAir11M2
+      case "iPad14,10", "iPad14,11": return iPadAir13M2
       case "iPad2,5", "iPad2,6", "iPad2,7": return iPadMini
       case "iPad4,4", "iPad4,5", "iPad4,6": return iPadMini2
       case "iPad4,7", "iPad4,8", "iPad4,9": return iPadMini3
@@ -586,6 +632,8 @@ public enum Device {
       case "iPad13,8", "iPad13,9", "iPad13,10", "iPad13,11": return iPadPro12Inch5
       case "iPad14,3", "iPad14,4": return iPadPro11Inch4
       case "iPad14,5", "iPad14,6": return iPadPro12Inch6
+      case "iPad16,3", "iPad16,4": return iPadPro11M4
+      case "iPad16,5", "iPad16,6": return iPadPro13M4
       case "AudioAccessory1,1": return homePod
       case "i386", "x86_64", "arm64": return simulator(mapToDevice(identifier: ProcessInfo().environment["SIMULATOR_MODEL_IDENTIFIER"] ?? "iOS"))
       default: return unknown(identifier)
@@ -627,12 +675,16 @@ public enum Device {
       case "Watch7,3": return appleWatchSeries9_41mm
       case "Watch7,4": return appleWatchSeries9_45mm
       case "Watch7,5": return appleWatchUltra2
+      case "Watch7,8", "Watch7,10": return appleWatchSeries10_42mm
+      case "Watch7,9", "Watch7,11": return appleWatchSeries10_46mm
       case "i386", "x86_64", "arm64": return simulator(mapToDevice(identifier: ProcessInfo().environment["SIMULATOR_MODEL_IDENTIFIER"] ?? "watchOS"))
       default: return unknown(identifier)
       }
     #elseif os(visionOS)
     // TODO: add proper visionOS support
     return .visionPro
+    #else
+      return unknown(identifier)
     #endif
   }
 
@@ -698,6 +750,10 @@ public enum Device {
         case .iPhone15Plus: return 6.7
         case .iPhone15Pro: return 6.1
         case .iPhone15ProMax: return 6.7
+        case .iPhone16: return 6.1
+        case .iPhone16Plus: return 6.7
+        case .iPhone16Pro: return 6.3
+        case .iPhone16ProMax: return 6.9
         case .iPad2: return 9.7
         case .iPad3: return 9.7
         case .iPad4: return 9.7
@@ -712,6 +768,8 @@ public enum Device {
         case .iPad10: return 10.9
         case .iPadAir4: return 10.9
         case .iPadAir5: return 10.9
+        case .iPadAir11M2: return 10.9
+        case .iPadAir13M2: return 12.9
         case .iPadMini: return 7.9
         case .iPadMini2: return 7.9
         case .iPadMini3: return 7.9
@@ -730,6 +788,8 @@ public enum Device {
         case .iPadPro12Inch5: return 12.9
         case .iPadPro11Inch4: return 11.0
         case .iPadPro12Inch6: return 12.9
+        case .iPadPro11M4: return 11.0
+        case .iPadPro13M4: return 13.0
         case .homePod: return -1
         case .simulator(let model): return model.diagonal
         case .unknown: return -1
@@ -762,6 +822,8 @@ public enum Device {
       case .appleWatchSeries9_41mm: return 1.9
       case .appleWatchSeries9_45mm: return 2.0
       case .appleWatchUltra2: return 2.2
+      case .appleWatchSeries10_42mm: return 1.9
+      case .appleWatchSeries10_46mm: return 2.0
       case .simulator(let model): return model.diagonal
       case .unknown: return -1
       }
@@ -815,6 +877,10 @@ public enum Device {
       case .iPhone15Plus: return (width: 9, height: 19.5)
       case .iPhone15Pro: return (width: 9, height: 19.5)
       case .iPhone15ProMax: return (width: 9, height: 19.5)
+      case .iPhone16: return (width: 9, height: 19.5)
+      case .iPhone16Plus: return (width: 9, height: 19.5)
+      case .iPhone16Pro: return (width: 9, height: 19.5)
+      case .iPhone16ProMax: return (width: 9, height: 19.5)
       case .iPad2: return (width: 3, height: 4)
       case .iPad3: return (width: 3, height: 4)
       case .iPad4: return (width: 3, height: 4)
@@ -829,6 +895,8 @@ public enum Device {
       case .iPad10: return (width: 41, height: 59)
       case .iPadAir4: return (width: 41, height: 59)
       case .iPadAir5: return (width: 41, height: 59)
+      case .iPadAir11M2: return (width: 59, height: 41)
+      case .iPadAir13M2: return (width: 683, height: 512)
       case .iPadMini: return (width: 3, height: 4)
       case .iPadMini2: return (width: 3, height: 4)
       case .iPadMini3: return (width: 3, height: 4)
@@ -847,6 +915,8 @@ public enum Device {
       case .iPadPro12Inch5: return (width: 512, height: 683)
       case .iPadPro11Inch4: return (width: 139, height: 199)
       case .iPadPro12Inch6: return (width: 512, height: 683)
+      case .iPadPro11M4: return (width: 605, height: 417)
+      case .iPadPro13M4: return (width: 4, height: 3)
       case .homePod: return (width: 4, height: 5)
       case .simulator(let model): return model.screenRatio
       case .unknown: return (width: -1, height: -1)
@@ -879,10 +949,14 @@ public enum Device {
       case .appleWatchSeries9_41mm: return (width: 4, height: 5)
       case .appleWatchSeries9_45mm: return (width: 4, height: 5)
       case .appleWatchUltra2: return (width: 4, height: 5)
+      case .appleWatchSeries10_42mm: return (width: 374, height: 446)
+      case .appleWatchSeries10_46mm: return (width: 416, height: 496)
       case .simulator(let model): return model.screenRatio
       case .unknown: return (width: -1, height: -1)
       }
     #elseif os(tvOS) || os(visionOS)
+      return (width: -1, height: -1)
+    #else
       return (width: -1, height: -1)
     #endif
   }
@@ -895,28 +969,28 @@ public enum Device {
 
     /// All iPhones
     public static var allPhones: [Device] {
-      return [.iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax]
+      return [.iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax]
     }
 
     /// All iPads
     public static var allPads: [Device] {
-      return [.iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPad9, .iPad10, .iPadAir4, .iPadAir5, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadMini6, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6]
+      return [.iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPad9, .iPad10, .iPadAir4, .iPadAir5, .iPadAir11M2, .iPadAir13M2, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadMini6, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6, .iPadPro11M4, .iPadPro13M4]
     }
 
     /// All X-Series Devices
     @available(*, deprecated, renamed: "allDevicesWithSensorHousing")
     public static var allXSeriesDevices: [Device] {
-      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax]
+      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax]
     }
 
     /// All Plus and Max-Sized Devices
     public static var allPlusSizedDevices: [Device] {
-      return [.iPhone6Plus, .iPhone6sPlus, .iPhone7Plus, .iPhone8Plus, .iPhoneXSMax, .iPhone11ProMax, .iPhone12ProMax, .iPhone13ProMax, .iPhone14Plus, .iPhone14ProMax, .iPhone15Plus, .iPhone15ProMax]
+      return [.iPhone6Plus, .iPhone6sPlus, .iPhone7Plus, .iPhone8Plus, .iPhoneXSMax, .iPhone11ProMax, .iPhone12ProMax, .iPhone13ProMax, .iPhone14Plus, .iPhone14ProMax, .iPhone15Plus, .iPhone15ProMax, .iPhone16Plus, .iPhone16ProMax]
     }
 
     /// All Pro Devices
     public static var allProDevices: [Device] {
-      return [.iPhone11Pro, .iPhone11ProMax, .iPhone12Pro, .iPhone12ProMax, .iPhone13Pro, .iPhone13ProMax, .iPhone14Pro, .iPhone14ProMax, .iPhone15Pro, .iPhone15ProMax, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6]
+      return [.iPhone11Pro, .iPhone11ProMax, .iPhone12Pro, .iPhone12ProMax, .iPhone13Pro, .iPhone13ProMax, .iPhone14Pro, .iPhone14ProMax, .iPhone15Pro, .iPhone15ProMax, .iPhone16Pro, .iPhone16ProMax, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6, .iPadPro11M4, .iPadPro13M4]
     }
 
     /// All mini Devices
@@ -996,22 +1070,21 @@ public enum Device {
       } else {
         return UIScreen.main.nativeScale > UIScreen.main.scale
       }
-      #endif
     }
 
     /// All Touch ID Capable Devices
     public static var allTouchIDCapableDevices: [Device] {
-      return [.iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneSE2, .iPhoneSE3, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPad9, .iPad10, .iPadAir4, .iPadAir5, .iPadMini3, .iPadMini4, .iPadMini5, .iPadMini6, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch]
+      return [.iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneSE2, .iPhoneSE3, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPad9, .iPad10, .iPadAir4, .iPadAir5, .iPadAir11M2, .iPadAir13M2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadMini6, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch]
     }
 
     /// All Face ID Capable Devices
     public static var allFaceIDCapableDevices: [Device] {
-      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6]
+      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6, .iPadPro11M4, .iPadPro13M4]
     }
 
     /// All Devices with Touch ID or Face ID
     public static var allBiometricAuthenticationCapableDevices: [Device] {
-      return [.iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPad9, .iPad10, .iPadAir4, .iPadAir5, .iPadMini3, .iPadMini4, .iPadMini5, .iPadMini6, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6]
+      return [.iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPad9, .iPad10, .iPadAir4, .iPadAir5, .iPadAir11M2, .iPadAir13M2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadMini6, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6, .iPadPro11M4, .iPadPro13M4]
     }
 
     /// Returns whether or not the device has Touch ID
@@ -1031,7 +1104,7 @@ public enum Device {
 
     /// All devices that feature a sensor housing in the screen
     public static var allDevicesWithSensorHousing: [Device] {
-      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax]
+      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax]
     }
 
     /// All simulator devices that feature a sensor housing in the screen
@@ -1046,7 +1119,7 @@ public enum Device {
 
     /// All devices that feature a screen with rounded corners.
     public static var allDevicesWithRoundedDisplayCorners: [Device] {
-      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPad10, .iPadAir4, .iPadAir5, .iPadMini6, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6]
+      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax, .iPad10, .iPadAir4, .iPadAir5, .iPadAir11M2, .iPadAir13M2, .iPadMini6, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6, .iPadPro11M4, .iPadPro13M4]
     }
 
     /// Returns whether or not the device has a screen with rounded corners.
@@ -1056,7 +1129,7 @@ public enum Device {
 
     /// All devices that have the Dynamic Island.
     public static var allDevicesWithDynamicIsland: [Device] {
-      return [.iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax]
+      return [.iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax]
     }
 
     /// Returns whether or not the device has the Dynamic Island.
@@ -1076,7 +1149,7 @@ public enum Device {
 
     /// All devices that support wireless charging.
     public static var allDevicesWithWirelessChargingSupport: [Device] {
-      return [.iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax]
+      return [.iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax]
     }
 
     /// Returns whether or not the device supports wireless charging.
@@ -1086,7 +1159,7 @@ public enum Device {
 
     /// All devices that support 5G.
     public static var allDevicesWith5gSupport: [Device] {
-      return [.iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPad10, .iPadAir5, .iPadMini6, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6]
+      return [.iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax, .iPad10, .iPadAir5, .iPadAir11M2, .iPadAir13M2, .iPadMini6, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6, .iPadPro11M4, .iPadPro13M4]
     }
 
     /// Returns whether or not the device has 5G support.
@@ -1096,7 +1169,7 @@ public enum Device {
 
     /// All devices that have a LiDAR sensor.
     public static var allDevicesWithALidarSensor: [Device] {
-      return [.iPhone12Pro, .iPhone12ProMax, .iPhone13Pro, .iPhone13ProMax, .iPhone14Pro, .iPhone14ProMax, .iPhone15Pro, .iPhone15ProMax, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6]
+      return [.iPhone12Pro, .iPhone12ProMax, .iPhone13Pro, .iPhone13ProMax, .iPhone14Pro, .iPhone14ProMax, .iPhone15Pro, .iPhone15ProMax, .iPhone16Pro, .iPhone16ProMax, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6, .iPadPro11M4, .iPadPro13M4]
     }
 
     /// Returns whether or not the device has a LiDAR sensor.
@@ -1106,7 +1179,7 @@ public enum Device {
 
     /// All devices that have a USB-C connectivity.
     public static var allDevicesWithUSBCConnectivity: [Device] {
-      return [.iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPad10, .iPadAir4, .iPadAir5, .iPadMini6, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6]
+      return [.iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax, .iPad10, .iPadAir4, .iPadAir5, .iPadAir11M2, .iPadAir13M2, .iPadMini6, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6, .iPadPro11M4, .iPadPro13M4]
     }
 
     /// Returns whether or not the device has a USB-C power supply.
@@ -1126,7 +1199,7 @@ public enum Device {
   #elseif os(watchOS)
     /// All Watches
     public static var allWatches: [Device] {
-       return [.appleWatchSeries0_38mm, .appleWatchSeries0_42mm, .appleWatchSeries1_38mm, .appleWatchSeries1_42mm, .appleWatchSeries2_38mm, .appleWatchSeries2_42mm, .appleWatchSeries3_38mm, .appleWatchSeries3_42mm, .appleWatchSeries4_40mm, .appleWatchSeries4_44mm, .appleWatchSeries5_40mm, .appleWatchSeries5_44mm, .appleWatchSeries6_40mm, .appleWatchSeries6_44mm, .appleWatchSE_40mm, .appleWatchSE_44mm, .appleWatchSeries7_41mm, .appleWatchSeries7_45mm, .appleWatchSeries8_41mm, .appleWatchSeries8_45mm, .appleWatchSE2_40mm, .appleWatchSE2_44mm, .appleWatchUltra, .appleWatchSeries9_41mm, .appleWatchSeries9_45mm, .appleWatchUltra2]
+       return [.appleWatchSeries0_38mm, .appleWatchSeries0_42mm, .appleWatchSeries1_38mm, .appleWatchSeries1_42mm, .appleWatchSeries2_38mm, .appleWatchSeries2_42mm, .appleWatchSeries3_38mm, .appleWatchSeries3_42mm, .appleWatchSeries4_40mm, .appleWatchSeries4_44mm, .appleWatchSeries5_40mm, .appleWatchSeries5_44mm, .appleWatchSeries6_40mm, .appleWatchSeries6_44mm, .appleWatchSE_40mm, .appleWatchSE_44mm, .appleWatchSeries7_41mm, .appleWatchSeries7_45mm, .appleWatchSeries8_41mm, .appleWatchSeries8_45mm, .appleWatchSE2_40mm, .appleWatchSE2_44mm, .appleWatchUltra, .appleWatchSeries9_41mm, .appleWatchSeries9_45mm, .appleWatchUltra2, .appleWatchSeries10_42mm, .appleWatchSeries10_46mm]
     }
 
     /// All simulator Watches
@@ -1165,6 +1238,8 @@ public enum Device {
       return allWatches
     #elseif os(visionOS)
       return [.visionPro]
+    #else
+      return []
     #endif
   }
 
@@ -1226,8 +1301,10 @@ public enum Device {
     guard isCurrent else { return nil }
     #if os(watchOS)
     return WKInterfaceDevice.current().name
-    #else
+    #elseif canImport(UIKit)
     return UIDevice.current.name
+    #else
+    return nil
     #endif
   }
 
@@ -1242,8 +1319,10 @@ public enum Device {
     } else {
       return UIDevice.current.systemName
     }
-    #else
+    #elseif canImport(UIKit)
     return UIDevice.current.systemName
+    #else
+    return nil
     #endif
   }
 
@@ -1252,8 +1331,10 @@ public enum Device {
     guard isCurrent else { return nil }
     #if os(watchOS)
     return WKInterfaceDevice.current().systemVersion
-    #else
+    #elseif canImport(UIKit)
     return UIDevice.current.systemVersion
+    #else
+    return nil
     #endif
   }
 
@@ -1262,8 +1343,10 @@ public enum Device {
     guard isCurrent else { return nil }
     #if os(watchOS)
     return WKInterfaceDevice.current().model
-    #else
+    #elseif canImport(UIKit)
     return UIDevice.current.model
+    #else
+    return nil
     #endif
   }
 
@@ -1272,8 +1355,10 @@ public enum Device {
     guard isCurrent else { return nil }
     #if os(watchOS)
     return WKInterfaceDevice.current().localizedModel
-    #else
+    #elseif canImport(UIKit)
     return UIDevice.current.localizedModel
+    #else
+    return nil
     #endif
   }
 
@@ -1316,13 +1401,17 @@ public enum Device {
       case .iPhone13ProMax: return 458
       case .iPhoneSE3: return 326
       case .iPhone14: return 460
-      case .iPhone14Plus: return 476
+      case .iPhone14Plus: return 458
       case .iPhone14Pro: return 460
       case .iPhone14ProMax: return 458
       case .iPhone15: return 460
-      case .iPhone15Plus: return 476
+      case .iPhone15Plus: return 460
       case .iPhone15Pro: return 460
       case .iPhone15ProMax: return 460
+      case .iPhone16: return 460
+      case .iPhone16Plus: return 460
+      case .iPhone16Pro: return 460
+      case .iPhone16ProMax: return 460
       case .iPad2: return 132
       case .iPad3: return 264
       case .iPad4: return 264
@@ -1337,6 +1426,8 @@ public enum Device {
       case .iPad10: return 264
       case .iPadAir4: return 264
       case .iPadAir5: return 264
+      case .iPadAir11M2: return 264
+      case .iPadAir13M2: return 264
       case .iPadMini: return 163
       case .iPadMini2: return 326
       case .iPadMini3: return 326
@@ -1355,6 +1446,8 @@ public enum Device {
       case .iPadPro12Inch5: return 264
       case .iPadPro11Inch4: return 264
       case .iPadPro12Inch6: return 264
+      case .iPadPro11M4: return 264
+      case .iPadPro13M4: return 264
       case .homePod: return -1
       case .simulator(let model): return model.ppi
       case .unknown: return nil
@@ -1387,9 +1480,16 @@ public enum Device {
     case .appleWatchSeries9_41mm: return 326
     case .appleWatchSeries9_45mm: return 326
     case .appleWatchUltra2: return 338
+    case .appleWatchSeries10_42mm: return 326
+    case .appleWatchSeries10_46mm: return 326
     case .simulator(let model): return model.ppi
     case .unknown: return nil
     }
+    #else
+    return nil
+    #elseif os(visionOS)
+    // TODO: Replace with proper implementation for visionOS.
+    return nil
     #else
     return nil
     #endif
@@ -1467,6 +1567,10 @@ extension Device: CustomStringConvertible {
       case .iPhone15Plus: return "iPhone 15 Plus"
       case .iPhone15Pro: return "iPhone 15 Pro"
       case .iPhone15ProMax: return "iPhone 15 Pro Max"
+      case .iPhone16: return "iPhone 16"
+      case .iPhone16Plus: return "iPhone 16 Plus"
+      case .iPhone16Pro: return "iPhone 16 Pro"
+      case .iPhone16ProMax: return "iPhone 16 Pro Max"
       case .iPad2: return "iPad 2"
       case .iPad3: return "iPad (3rd generation)"
       case .iPad4: return "iPad (4th generation)"
@@ -1481,6 +1585,8 @@ extension Device: CustomStringConvertible {
       case .iPad10: return "iPad (10th generation)"
       case .iPadAir4: return "iPad Air (4th generation)"
       case .iPadAir5: return "iPad Air (5th generation)"
+      case .iPadAir11M2: return "iPad Air (11-inch) (M2)"
+      case .iPadAir13M2: return "iPad Air (13-inch) (M2)"
       case .iPadMini: return "iPad Mini"
       case .iPadMini2: return "iPad Mini 2"
       case .iPadMini3: return "iPad Mini 3"
@@ -1499,6 +1605,8 @@ extension Device: CustomStringConvertible {
       case .iPadPro12Inch5: return "iPad Pro (12.9-inch) (5th generation)"
       case .iPadPro11Inch4: return "iPad Pro (11-inch) (4th generation)"
       case .iPadPro12Inch6: return "iPad Pro (12.9-inch) (6th generation)"
+      case .iPadPro11M4: return "iPad Pro (11-inch) (M4)"
+      case .iPadPro13M4: return "iPad Pro (13-inch) (M4)"
       case .homePod: return "HomePod"
       case .simulator(let model): return "Simulator (\(model.description))"
       case .unknown(let identifier): return identifier
@@ -1531,6 +1639,8 @@ extension Device: CustomStringConvertible {
       case .appleWatchSeries9_41mm: return "Apple Watch Series 9 41mm"
       case .appleWatchSeries9_45mm: return "Apple Watch Series 9 45mm"
       case .appleWatchUltra2: return "Apple Watch Ultra2"
+      case .appleWatchSeries10_42mm: return "Apple Watch Series 10 42mm"
+      case .appleWatchSeries10_46mm: return "Apple Watch Series 10 46mm"
       case .simulator(let model): return "Simulator (\(model.description))"
       case .unknown(let identifier): return identifier
       }
@@ -1549,6 +1659,11 @@ extension Device: CustomStringConvertible {
       case .simulator(let model): return "Simulator (\(model.description))"
       case .unknown(let identifier): return identifier
       }    
+    #else
+      switch self {
+      case .simulator(let model): return "Simulator (\(model.description))"
+      case .unknown(let identifier): return identifier
+      }
     #endif
   }
 
@@ -1601,6 +1716,10 @@ extension Device: CustomStringConvertible {
       case .iPhone15Plus: return "iPhone 15 Plus"
       case .iPhone15Pro: return "iPhone 15 Pro"
       case .iPhone15ProMax: return "iPhone 15 Pro Max"
+      case .iPhone16: return "iPhone 16"
+      case .iPhone16Plus: return "iPhone 16 Plus"
+      case .iPhone16Pro: return "iPhone 16 Pro"
+      case .iPhone16ProMax: return "iPhone 16 Pro Max"
       case .iPad2: return "iPad 2"
       case .iPad3: return "iPad (3rd generation)"
       case .iPad4: return "iPad (4th generation)"
@@ -1615,6 +1734,8 @@ extension Device: CustomStringConvertible {
       case .iPad10: return "iPad (10th generation)"
       case .iPadAir4: return "iPad Air (4th generation)"
       case .iPadAir5: return "iPad Air (5th generation)"
+      case .iPadAir11M2: return "iPad Air (11-inch) (M2)"
+      case .iPadAir13M2: return "iPad Air (13-inch) (M2)"
       case .iPadMini: return "iPad Mini"
       case .iPadMini2: return "iPad Mini 2"
       case .iPadMini3: return "iPad Mini 3"
@@ -1633,6 +1754,8 @@ extension Device: CustomStringConvertible {
       case .iPadPro12Inch5: return "iPad Pro (12.9-inch) (5th generation)"
       case .iPadPro11Inch4: return "iPad Pro (11-inch) (4th generation)"
       case .iPadPro12Inch6: return "iPad Pro (12.9-inch) (6th generation)"
+      case .iPadPro11M4: return "iPad Pro (11-inch) (M4)"
+      case .iPadPro13M4: return "iPad Pro (13-inch) (M4)"
       case .homePod: return "HomePod"
       case .simulator(let model): return "Simulator (\(model.safeDescription))"
       case .unknown(let identifier): return identifier
@@ -1665,6 +1788,8 @@ extension Device: CustomStringConvertible {
       case .appleWatchSeries9_41mm: return "Apple Watch Series 9 41mm"
       case .appleWatchSeries9_45mm: return "Apple Watch Series 9 45mm"
       case .appleWatchUltra2: return "Apple Watch Ultra2"
+      case .appleWatchSeries10_42mm: return "Apple Watch Series 10 42mm"
+      case .appleWatchSeries10_46mm: return "Apple Watch Series 10 46mm"
       case .simulator(let model): return "Simulator (\(model.safeDescription))"
       case .unknown(let identifier): return identifier
       }
@@ -1680,6 +1805,8 @@ extension Device: CustomStringConvertible {
     #elseif os(visionOS)
       switch self {
       case .visionPro: return "Apple Vision Pro"
+    #else
+      switch self {
       case .simulator(let model): return "Simulator (\(model.safeDescription))"
       case .unknown(let identifier): return identifier
       }
@@ -1840,17 +1967,21 @@ extension Device {
       This enum describes the state of the orientation.
       - Landscape: The device is in Landscape Orientation
       - Portrait:  The device is in Portrait Orientation
+      - Unknown:   The device orientation is unknown.
     */
     public enum Orientation {
       case landscape
       case portrait
+      case unknown
     }
 
     public var orientation: Orientation {
       if UIDevice.current.orientation.isLandscape {
         return .landscape
-      } else {
+      } else if UIDevice.current.orientation.isPortrait {
         return .portrait
+      } else {
+        return .unknown
       }
     }
 }
@@ -1918,8 +2049,10 @@ extension Device {
 
   /**
     This option set describes the current Apple Pencils
-    - firstGeneration:  1st Generation Apple Pencil
+    - firstGeneration: 1st Generation Apple Pencil
     - secondGeneration: 2nd Generation Apple Pencil
+    - firstGenerationUsbC: 1st Generation Apple Pencil (USB-C)
+    - pro: Apple Pencil Pro
    */
   public struct ApplePencilSupport: OptionSet {
 
@@ -1931,11 +2064,13 @@ extension Device {
 
     public static let firstGeneration = ApplePencilSupport(rawValue: 0x01)
     public static let secondGeneration = ApplePencilSupport(rawValue: 0x02)
+    public static let firstGenerationUsbC = ApplePencilSupport(rawValue: 0x03)
+    public static let pro = ApplePencilSupport(rawValue: 0x04)
   }
 
   /// All Apple Pencil Capable Devices
   public static var allApplePencilCapableDevices: [Device] {
-    return [.iPad6, .iPadAir3, .iPad7, .iPad8, .iPad9, .iPad10, .iPadAir4, .iPadAir5, .iPadMini5, .iPadMini6, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6]
+    return [.iPad6, .iPadAir3, .iPad7, .iPad8, .iPad9, .iPad10, .iPadAir4, .iPadAir5, .iPadAir11M2, .iPadAir13M2, .iPadMini5, .iPadMini6, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6, .iPadPro11M4, .iPadPro13M4]
   }
 
   /// Returns supported version of the Apple Pencil
@@ -1946,23 +2081,27 @@ extension Device {
       case .iPad7: return .firstGeneration
       case .iPad8: return .firstGeneration
       case .iPad9: return .firstGeneration
-      case .iPad10: return .firstGeneration
       case .iPadMini5: return .firstGeneration
       case .iPadPro9Inch: return .firstGeneration
       case .iPadPro12Inch: return .firstGeneration
       case .iPadPro12Inch2: return .firstGeneration
       case .iPadPro10Inch: return .firstGeneration
-      case .iPadAir4: return .secondGeneration
-      case .iPadAir5: return .secondGeneration
-      case .iPadMini6: return .secondGeneration
-      case .iPadPro11Inch: return .secondGeneration
-      case .iPadPro12Inch3: return .secondGeneration
-      case .iPadPro11Inch2: return .secondGeneration
-      case .iPadPro12Inch4: return .secondGeneration
-      case .iPadPro11Inch3: return .secondGeneration
-      case .iPadPro12Inch5: return .secondGeneration
-      case .iPadPro11Inch4: return .secondGeneration
-      case .iPadPro12Inch6: return .secondGeneration
+      case .iPad10: return .firstGenerationUsbC
+      case .iPadAir4: return [.secondGeneration, .firstGenerationUsbC]
+      case .iPadAir5: return [.secondGeneration, .firstGenerationUsbC]
+      case .iPadMini6: return [.secondGeneration, .firstGenerationUsbC]
+      case .iPadPro11Inch: return [.secondGeneration, .firstGenerationUsbC]
+      case .iPadPro12Inch3: return [.secondGeneration, .firstGenerationUsbC]
+      case .iPadPro11Inch2: return [.secondGeneration, .firstGenerationUsbC]
+      case .iPadPro12Inch4: return [.secondGeneration, .firstGenerationUsbC]
+      case .iPadPro11Inch3: return [.secondGeneration, .firstGenerationUsbC]
+      case .iPadPro12Inch5: return [.secondGeneration, .firstGenerationUsbC]
+      case .iPadPro11Inch4: return [.secondGeneration, .firstGenerationUsbC]
+      case .iPadPro12Inch6: return [.secondGeneration, .firstGenerationUsbC]
+      case .iPadPro11M4: return [.secondGeneration, .firstGenerationUsbC, .pro]
+      case .iPadPro13M4: return [.secondGeneration, .firstGenerationUsbC, .pro]
+      case .iPadAir11M2: return [.secondGeneration, .pro]
+      case .iPadAir13M2: return [.secondGeneration, .pro]
       case .simulator(let model): return model.applePencilSupport
       default: return []
     }
@@ -2018,6 +2157,8 @@ extension Device {
       case .iPad10: return [.wide]
       case .iPadAir4: return [.wide]
       case .iPadAir5: return [.wide]
+      case .iPadAir11M2: return [.wide]
+      case .iPadAir13M2: return [.wide]
       case .iPadMini: return [.wide]
       case .iPadMini2: return [.wide]
       case .iPadMini3: return [.wide]
@@ -2030,6 +2171,8 @@ extension Device {
       case .iPadPro10Inch: return [.wide]
       case .iPadPro11Inch: return [.wide]
       case .iPadPro12Inch3: return [.wide]
+      case .iPadPro11M4: return [.wide]
+      case .iPadPro13M4: return [.wide]
       case .iPhone7Plus: return [.wide, .telephoto]
       case .iPhone8Plus: return [.wide, .telephoto]
       case .iPhoneX: return [.wide, .telephoto]
@@ -2044,6 +2187,8 @@ extension Device {
       case .iPhone14Plus: return [.wide, .ultraWide]
       case .iPhone15: return [.wide, .ultraWide]
       case .iPhone15Plus: return [.wide, .ultraWide]
+      case .iPhone16: return [.wide, .ultraWide]
+      case .iPhone16Plus: return [.wide, .ultraWide]
       case .iPadPro11Inch2: return [.wide, .ultraWide]
       case .iPadPro12Inch4: return [.wide, .ultraWide]
       case .iPadPro11Inch3: return [.wide, .ultraWide]
@@ -2060,13 +2205,15 @@ extension Device {
       case .iPhone14ProMax: return [.wide, .telephoto, .ultraWide]
       case .iPhone15Pro: return [.wide, .telephoto, .ultraWide]
       case .iPhone15ProMax: return [.wide, .telephoto, .ultraWide]
+      case .iPhone16Pro: return [.wide, .telephoto, .ultraWide]
+      case .iPhone16ProMax: return [.wide, .telephoto, .ultraWide]
       default: return []
     }
   }
 
   /// All devices that feature a camera
   public static var allDevicesWithCamera: [Device] {
-    return [.iPodTouch5, .iPodTouch6, .iPodTouch7, .iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPad9, .iPad10, .iPadAir4, .iPadAir5, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadMini6, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6]
+    return [.iPodTouch5, .iPodTouch6, .iPodTouch7, .iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax, .iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPad9, .iPad10, .iPadAir4, .iPadAir5, .iPadAir11M2, .iPadAir13M2, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadMini6, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6, .iPadPro11M4, .iPadPro13M4]
   }
 
   /// All devices that feature a normal camera
@@ -2077,17 +2224,17 @@ extension Device {
 
   /// All devices that feature a wide camera
   public static var allDevicesWithWideCamera: [Device] {
-    return [.iPodTouch5, .iPodTouch6, .iPodTouch7, .iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPad9, .iPad10, .iPadAir4, .iPadAir5, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadMini6, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6]
+    return [.iPodTouch5, .iPodTouch6, .iPodTouch7, .iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhoneSE3, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax, .iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPad9, .iPad10, .iPadAir4, .iPadAir5, .iPadAir11M2, .iPadAir13M2, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadMini6, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6, .iPadPro11M4, .iPadPro13M4]
   }
 
   /// All devices that feature a telephoto camera
   public static var allDevicesWithTelephotoCamera: [Device] {
-    return [.iPhone7Plus, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhone11Pro, .iPhone11ProMax, .iPhone12Pro, .iPhone12ProMax, .iPhone13Pro, .iPhone13ProMax, .iPhone14Pro, .iPhone14ProMax, .iPhone15Pro, .iPhone15ProMax]
+    return [.iPhone7Plus, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhone11Pro, .iPhone11ProMax, .iPhone12Pro, .iPhone12ProMax, .iPhone13Pro, .iPhone13ProMax, .iPhone14Pro, .iPhone14ProMax, .iPhone15Pro, .iPhone15ProMax, .iPhone16Pro, .iPhone16ProMax]
   }
 
   /// All devices that feature an ultra wide camera
   public static var allDevicesWithUltraWideCamera: [Device] {
-    return [.iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6]
+    return [.iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhone12, .iPhone12Mini, .iPhone12Pro, .iPhone12ProMax, .iPhone13, .iPhone13Mini, .iPhone13Pro, .iPhone13ProMax, .iPhone14, .iPhone14Plus, .iPhone14Pro, .iPhone14ProMax, .iPhone15, .iPhone15Plus, .iPhone15Pro, .iPhone15ProMax, .iPhone16, .iPhone16Plus, .iPhone16Pro, .iPhone16ProMax, .iPadPro11Inch2, .iPadPro12Inch4, .iPadPro11Inch3, .iPadPro12Inch5, .iPadPro11Inch4, .iPadPro12Inch6]
   }
 
   /// Returns whether or not the current device has a camera
@@ -2178,8 +2325,12 @@ extension Device {
     case a15Bionic
     case a16Bionic
     case a17Pro
+    case a18
+    case a18Pro
     case m1
     case m2
+    case m3
+    case m4
   #elseif os(watchOS)
     case s1
     case s1P
@@ -2191,6 +2342,7 @@ extension Device {
     case s7
     case s8
     case s9
+    case s10
   #endif
     case unknown
   }
@@ -2241,6 +2393,10 @@ extension Device {
       case .iPhone15Plus: return .a16Bionic
       case .iPhone15Pro: return .a17Pro
       case .iPhone15ProMax: return .a17Pro
+      case .iPhone16: return .a18
+      case .iPhone16Plus: return .a18
+      case .iPhone16Pro: return .a18Pro
+      case .iPhone16ProMax: return .a18Pro
       case .iPad2: return .a5
       case .iPad3: return .a5X
       case .iPad4: return .a6X
@@ -2255,6 +2411,8 @@ extension Device {
       case .iPad10: return .a14Bionic
       case .iPadAir4: return .a14Bionic
       case .iPadAir5: return .m1
+      case .iPadAir11M2: return .m2
+      case .iPadAir13M2: return .m2
       case .iPadMini: return .a5
       case .iPadMini2: return .a7
       case .iPadMini3: return .a7
@@ -2273,6 +2431,8 @@ extension Device {
       case .iPadPro12Inch5: return .m1
       case .iPadPro11Inch4: return .m2
       case .iPadPro12Inch6: return .m2
+      case .iPadPro11M4: return .m4
+      case .iPadPro13M4: return .m4
       case .homePod: return .a8
       case .simulator(let model): return model.cpu
       case .unknown: return .unknown
@@ -2305,6 +2465,8 @@ extension Device {
       case .appleWatchSeries9_41mm: return .s9
       case .appleWatchSeries9_45mm: return .s9
       case .appleWatchUltra2: return .s9
+      case .appleWatchSeries10_42mm: return .s10
+      case .appleWatchSeries10_46mm: return .s10
       case .simulator(let model): return model.cpu
       case .unknown: return .unknown
     }
@@ -2317,6 +2479,9 @@ extension Device {
       case .simulator(let model): return model.cpu
       case .unknown: return .unknown
     }
+  #elseif os(visionOS)
+    // TODO: Replace with proper implementation for visionOS.
+    return .unknown
   #else
     return .unknown
   #endif
@@ -2350,8 +2515,12 @@ extension Device.CPU: CustomStringConvertible {
       case .a15Bionic: return "A15 Bionic"
       case .a16Bionic: return "A16 Bionic"
       case .a17Pro: return "A17 Pro"
+      case .a18: return "A18"
+      case .a18Pro: return "A18 Pro"
       case .m1: return "M1"
       case .m2: return "M2"
+      case .m3: return "M3"
+      case .m4: return "M4"
       case .unknown: return "unknown"
     }
   #elseif os(watchOS)
@@ -2366,8 +2535,12 @@ extension Device.CPU: CustomStringConvertible {
       case .s7: return "S7"
       case .s8: return "S8"
       case .s9: return "S9"
+      case .s10: return "S10"
       case .unknown: return "unknown"
     }
+  #elseif os(visionOS)
+    // TODO: Replace with proper implementation for visionOS.
+    return "unknown"
   #else
     return "unknown"
   #endif