+++ Tests/Tests.swift
@@ -17,20 +17,31 @@ class DeviceKitTests: XCTestCase {
   let device = Device.current
 
   func testDeviceSimulator() {
+    #if os(macOS)
+    XCTAssertFalse(device.isOneOf(Device.allSimulators))
+    #else
     XCTAssertTrue(device.isOneOf(Device.allSimulators))
+    #endif
   }
 
   func testIsSimulator() {
+    #if os(macOS)
+    XCTAssertFalse(device.isSimulator)
+    #else
     XCTAssertTrue(device.isSimulator)
+    #endif
   }
 
   func testDeviceDescription() {
+    #if os(macOS)
+    #else
     XCTAssertTrue(device.description.hasPrefix("Simulator"))
     XCTAssertTrue(device.description.contains("iPhone")
       || device.description.contains("iPad")
       || device.description.contains("iPod")
       || device.description.contains("TV")
       || device.description.contains("Apple Watch"))
+    #endif
   }
 
   func testIsCanvas() {
@@ -40,6 +51,8 @@ class DeviceKitTests: XCTestCase {
     let otherDevice: Device = device == .appleTVHD ? .appleTV4K : .appleTVHD
     #elseif os(watchOS)
     let otherDevice: Device = device == .appleWatchUltra ? .appleWatchSeries8_41mm : .appleWatchUltra
+    #else
+    let otherDevice: Device = .unknown("mac")
     #endif
     XCTAssertEqual(otherDevice.isCanvas, nil)
     XCTAssertEqual(device.isCanvas, false)
@@ -442,7 +455,9 @@ class DeviceKitTests: XCTestCase {
       .iPhone14Plus,
       .iPhone14ProMax,
       .iPhone15Plus,
-      .iPhone15ProMax
+      .iPhone15ProMax,
+      .iPhone16Plus,
+      .iPhone16ProMax,
     ])
   }
 
@@ -458,6 +473,8 @@ class DeviceKitTests: XCTestCase {
       .iPhone14ProMax,
       .iPhone15Pro,
       .iPhone15ProMax,
+      .iPhone16Pro,
+      .iPhone16ProMax,
       .iPadPro9Inch,
       .iPadPro12Inch,
       .iPadPro12Inch2,
@@ -469,7 +486,9 @@ class DeviceKitTests: XCTestCase {
       .iPadPro11Inch3,
       .iPadPro12Inch5,
       .iPadPro11Inch4,
-      .iPadPro12Inch6
+      .iPadPro12Inch6,
+      .iPadPro11M4,
+      .iPadPro13M4,
     ])
   }
 
@@ -496,7 +515,11 @@ class DeviceKitTests: XCTestCase {
       .iPhone15,
       .iPhone15Plus,
       .iPhone15Pro,
-      .iPhone15ProMax
+      .iPhone15ProMax,
+      .iPhone16,
+      .iPhone16Plus,
+      .iPhone16Pro,
+      .iPhone16ProMax,
     ]
     for device in Device.allRealDevices {
       XCTAssertTrue(device.hasDynamicIsland == device.isOneOf(dynamicIslandDevices), "testHasDynamicIsland failed for \(device.description)")
@@ -522,13 +545,21 @@ class DeviceKitTests: XCTestCase {
       .iPhone15Plus,
       .iPhone15Pro,
       .iPhone15ProMax,
+      .iPhone16,
+      .iPhone16Plus,
+      .iPhone16Pro,
+      .iPhone16ProMax,
       .iPad10,
       .iPadAir5,
+      .iPadAir11M2,
+      .iPadAir13M2,
       .iPadMini6,
       .iPadPro11Inch3,
       .iPadPro12Inch5,
       .iPadPro11Inch4,
-      .iPadPro12Inch6
+      .iPadPro12Inch6,
+      .iPadPro11M4,
+      .iPadPro13M4,
     ]
     for device in Device.allRealDevices {
       XCTAssertTrue(device.has5gSupport == device.isOneOf(has5gDevices), "testHasDynamicIsland failed for \(device.description)")
@@ -605,12 +636,16 @@ class DeviceKitTests: XCTestCase {
       .iPhone14ProMax,
       .iPhone15Pro,
       .iPhone15ProMax,
+      .iPhone16Pro,
+      .iPhone16ProMax,
       .iPadPro11Inch2,
       .iPadPro12Inch4,
       .iPadPro11Inch3,
       .iPadPro12Inch5,
       .iPadPro11Inch4,
-      .iPadPro12Inch6
+      .iPadPro12Inch6,
+      .iPadPro11M4,
+      .iPadPro13M4,
     ]
     for device in Device.allRealDevices {
       XCTAssertTrue(device.hasLidarSensor == device.isOneOf(lidarDevices), "testLidarValues failed for \(device.description)")
@@ -623,9 +658,15 @@ class DeviceKitTests: XCTestCase {
       .iPhone15Plus,
       .iPhone15Pro,
       .iPhone15ProMax,
+      .iPhone16,
+      .iPhone16Plus,
+      .iPhone16Pro,
+      .iPhone16ProMax,
       .iPad10,
       .iPadAir4,
       .iPadAir5,
+      .iPadAir11M2,
+      .iPadAir13M2,
       .iPadMini6,
       .iPadPro11Inch,
       .iPadPro12Inch3,
@@ -634,7 +675,9 @@ class DeviceKitTests: XCTestCase {
       .iPadPro11Inch3,
       .iPadPro12Inch5,
       .iPadPro11Inch4,
-      .iPadPro12Inch6
+      .iPadPro12Inch6,
+      .iPadPro11M4,
+      .iPadPro13M4,
     ]
     for device in Device.allRealDevices {
       XCTAssertTrue(device.hasUSBCConnectivity == device.isOneOf(usbCDevices), "testHasUSBCConnectivity failed for \(device.description)")